+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.0.0
+
+* Added `NullSafeWrap` widget.
+* Renamed parent builders as `builder` and `fallback`.
+* Removed `context` from builders to provide simplicity.
+
 ## 0.0.1
 
 * Initial release.